render.js 파일분리 (20.11.09)
list랑 aside파일을 만들고 render.js의 list관련 함수만 파일 분리 해놨습니다.
