+++ TeamProject/static/js/auth.js
@@ -1,5 +1,5 @@
 import * as URL from "./config.js"
-import { signup_FetchAPI, get_userinfo_FetchAPI,login_FetchAPI } from './post/fetch.js'
+import { signup_FetchAPI, get_userinfo_FetchAPI,login_FetchAPI } from './board/fetch.js'
 
 // --------- 접속 시 실행 ------------
 before_login();
