+++ TeamProject/templates/post.html
@@ -83,11 +83,12 @@
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
 
 <script type="module" src="/static/js/config.js"></script>
-<script type="module" src="/static/js/post/event.js"></script>
-<script type="module" src="/static/js/post/main.js"></script>
-<script type="module" src="/static/js/post/render.js"></script>
-<script type="module" src="/static/js/post/fetch.js"></script>
-<script type="module" src="/static/js/post/router.js"></script>
+<script type="module" src="/static/js/board/event.js"></script>
+<script type="module" src="/static/js/board/main.js"></script>
+<script type="module" src="/static/js/board/render.js"></script>
+<script type="module" src="/static/js/board/fetch.js"></script>
+<script type="module" src="/static/js/board/router.js"></script>
+<script type="module" src="/static/js/board/list/render.js"></script>
 
 <script type="module" src="/static/js/auth.js"></script>
